refactor(LC009): split missing-asnotracking analysis helpers
- separate query-chain analysis and write-detection logic
- keep analyzer behavior unchanged while reducing local complexity
